Refactor to reuse functions from Simple.hs
